fix: renamed component for readability
